+++ testcontainers-advanced-imagebuilder/src/main/java/software/xdev/testcontainers/imagebuilder/transfer/DefaultTransferFilesCreator.java
@@ -82,7 +82,8 @@ public class DefaultTransferFilesCreator implements TransferFilesCreator
 		final Set<String> preGitIgnoreLines,
 		final Predicate<String> ignoreFileLineFilter,
 		final Set<String> postGitIgnoreLines,
-		final Set<String> alwaysIncludedRelativePaths)
+		final Set<String> alwaysIncludedRelativePaths,
+		final boolean useWinNTFSJunctionFix)
 	{
 		try
 		{
@@ -100,7 +101,8 @@ public class DefaultTransferFilesCreator implements TransferFilesCreator
 			
 			return this.walkFilesAndDetermineTransfer(
 				ignoreNode,
-				alwaysIncludedRelativePaths);
+				alwaysIncludedRelativePaths,
+				useWinNTFSJunctionFix);
 		}
 		catch(final IOException ioe)
 		{
@@ -133,9 +135,11 @@ public class DefaultTransferFilesCreator implements TransferFilesCreator
 	
 	protected Map<Path, String> walkFilesAndDetermineTransfer(
 		final IgnoreNode ignoreNode,
-		final Set<String> alwaysIncludedRelativePaths) throws IOException
+		final Set<String> alwaysIncludedRelativePaths,
+		final boolean useWinNTFSJunctionFix) throws IOException
 	{
 		try(final Stream<Path> walk = findFiles(
+			useWinNTFSJunctionFix,
 			this.baseDir,
 			Integer.MAX_VALUE,
 			// Ignore directories
@@ -161,13 +165,14 @@ public class DefaultTransferFilesCreator implements TransferFilesCreator
 	}
 	
 	protected static Stream<Path> findFiles(
+		final boolean useWinNTFSJunctionFix,
 		final Path start,
 		final int maxDepth,
 		final BiPredicate<Path, BasicFileAttributes> matcher,
 		final FileVisitOption... options)
 		throws IOException
 	{
-		return WinNTFSJunctionFiles.shouldBeApplied(start)
+		return useWinNTFSJunctionFix && WinNTFSJunctionFiles.shouldBeApplied(start)
 			? WinNTFSJunctionFiles.find(start, maxDepth, matcher, options)
 			: Files.find(start, maxDepth, matcher, options);
 	}