+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.2.0
+* Added an explicit option for enabling the Windows NTFS junction fix: ``withUseWinNTFSJunctionFix`` #155
+  * Enabling it also requires adding ``--add-exports java.base/sun.nio.fs=ALL-UNNAMED`` or performance will be impacted by 100x due to non-accessible file attributes cache
+  * This option is temporary and will be removed once the underlying JDK bug was fixed
+* The default logger of ``AdvancedImageFromDockerFile`` now also includes ``dockerImageName`` to make it easier to distinguish between parallel builds
+
 # 2.1.1
 * Addresses a JDK bug which results in a crash or "infinite" loop when encountering recursive NTFS junctions on Windows #155